train_test

+++ datasets/preprocessing/DeepLIIF_preprocessing.py
@@ -42,9 +42,9 @@ if __name__ == '__main__':
             for i in range(len(img_classes)):
                 crop_img = img[:, i*img_size: (i+1)*img_size, :]
                 if i == 5:
-                    positive = crop_img[:, :, 0] == 1
-                    negative = crop_img[:, :, 2] == 0
-                    instance = np.sum(crop_img, axis=2)>0
+                    positive = (crop_img[:, :, 0] == 1).astype(np.uint8)
+                    negative = (crop_img[:, :, 2] == 0).astype(np.uint8)
+                    instance = (np.sum(crop_img, axis=2)>0).astype(np.uint8)
                     instance = make_instance_sonnet(instance, positive+negative)
 
                     positive = Image.fromarray(positive.astype('uint8'))
